Fix some bugs, add docs for udptl, named_acl, and app_confbridge.
The format of descriptions via 'config show' needs to be re-worked to be similar
to how 'core show function' and the like work. Right now I'm just trying to get
some info in to start with so that there won't be horrible errors strewn across
the screen due to lack of documentation. Next up, res_xmpp and chan_motif.