Test for Asterisk Version info
Doxygen uses the ASTERISKVERSION as a sub header. If a SVN export is done and no .svn or .version file exists it defualts to UNKNOWN__and_probably_unsupported which is honest but not great for the online docs.  During the "make progdocs" I added a test for this and just warned and ommitted the version.

(issue ASTERISK-20259)