Create a threadpool API for Asterisk.
A common theme for a lot of feature development has been that
a thread pool would be useful.

In this branch, I plan to implement a generic thread pool whose
policies can be controlled by listeners who react to activity in the pool.

This will more-or-less be a C implementation of the thread pool implemented
in Asterisk SCF.

The first victim^W entity to use the thread pool will be the PBX. It currently
allocates a new thread for each new call and tears it down when the call is
completed. Using a thread pool could be useful for systems that have a steady
amount of traffic on them.

This may, just *may*, also find use in the new chan_sip that is being developed.