This is a branch intended to fix a potential memory leak present in taskprocessors.
See https://issues.asterisk.org/jira/browse/ASTERISK-20786

Since there are quite a few changes to be made towards fixing this, I'm doing the
work in a developer branch. This branch is based off Asterisk 1.8, since the issue
is present there. The idea is that the brunt of the work will need to be done in
1.8, and that incremental changes in future branches can easily be made when this
work gets merged.

I expect also that there could be objection to this being merged into 1.8, but I'll
cross that bridge when the time comes. Either way, the work here will be useful.