Add some commentary to stub functions to explain some things.
For the function to fill in an ast_party_id, the structure of the
function has been outlined, but it has not been filled in because
at the moment it is pretty low-priority.

Explanations behind the PJSIP module load/unload/start/stop functions
have been made, specifically about why we don't actually do anything
in those callbacks.