Add threadpool support to res_sip.
Here's how it works:

An ast_sip_work contains a taskprocessor. When a session
wishes to do work, it queues a task onto this taskprocessor.
When the taskprocessor listener is told that a task has been pushed,
it reacts by enqueueing itself onto the threadpool. The threadpool's
threads then pull tasks out of the ast_sip_work's taskprocessor
until it's empty.

What this does is make all tasks pertaining to a session get executed
sequentially by whatever thread is currently available in the threadpool.

The threadpool itself has hardcoded options in it for now. This will be
controllable via general settings for res_sip.