Add an endpoint identifier that matches based on IP address.
This adds a primitive version of endpoint location in order to
get an IP address for an endpoint and to get an endpoint based
on an IP address.

Some potential improvements:
1) The sorcery object for retrieving an endpoint by IP address
may be better suited to living in res_sip_endpoint_identifier_ip.c
since it is unknown whether other services will find this sort of
thing useful.
2) When you are looking for the location of a particular endpoint,
getting an IP address has limited use. Getting A SIP URI or some
other sort of FQDN would be much more useful.
3) There is no method at the moment to set endpoint location
information aside from when it is read from the configuration
file. Such a method will be necessary when it comes time to
write a registrar

Still, this means that we can now recognize incoming calls by
the source IP address, which is pretty cool.