Add a proof of concept for access control lists for SIP.
This uses sorcery in order to create access control lists
for SIP communications. It runs at a very high priority so
that it can reject messages early in the message-handling
process.

This is written as a proof of concept by in my spare time
rather than as something that is likely to be in the final
version of res_sip-based code. I expect that at least some
of this could be used as a basis for the real deal, but
there are some drawbacks to the code as it exists:

1) We have to contend with the fact that multiple global ACL objects may
be configured. It doesn't really make sense to allow for this, but
I'm not sure how to tell sorcery to only allow a single instance
of an object type to exist.
2) There is no support for per-endpoint ACLs, yet.

This is *untested*. I just cobbled this together this evening. I'll
give it some testing in the upcoming days to make sure everything
is sane.