Allow for pushing tasks with no associated work.
For tasks that have no direct association with other tasks, there
is no need of creating the overhead of an ast_sip_work. Simply allow
the task to be pushed directly onto the threadpool instead.

This is being used now for certain startup and shutdown operations, and
you'll soon see that it will be used for OPTIONS handling as well.
Stay tuned!