Make OPTIONS handling use the threadpool.
For those wondering about the changes with regards to dialogs in this commit,
pjsip_rx_data_clone() does not copy the mod_data of the rdata. This means that
function calls like pjsip_rdata_get_dlg() and pjsip_rdata_get_tsx() will not
work on a cloned rdata structure. The solution here is to save the information
along with the rdata when we pass the data to the threadpool.