Add a backport from PJSIP trunk regarding authentication lookup.
Like with the pjsip_rx_data_clone() change previously, this is a backport
of functionality in PJSIP trunk that is due for release in PJSIP 2.0.5.
For ease of development, the change is being made here and we will maintain
the changes here and in the external PJSIP location until PJSIP 2.0.5 is
released.