Massive changes to authentication API.
The authenticator callbacks have been changed. The authentication_required()
callback remains, but the other two have been replaced with a single general
check_authentication() callback. This gives more freedom to the authenticator
to authenticate in the fashion that fits best. The old method was awkwardly
forcing digest authentication on the authenticator, or forcing them to put
a square peg in a round hole in order to use a different scheme.

All digest-related authentication lives locally in res_sip_authenticator
(that will likely see a name change soon).

res_sip_authenticator also has some hardcoded data in order to do some testing.
And hey, authentication worked properly when presented with the correct
credentials!

I also changed the storage method of the registered SIP authenticator, mainly
because I was completely misusing an ao2_global_obj. It can potentially change
back to using a global object, but for now, just trying to get something working,
I'm keeping it simple.