Construct a temporary way of using the endpoint to determine if authentication succeeds.
The "secret" endpoint configuration option is the only one added at this point. Configuration
of realm, authentication username, and potential other methods of authentication should be
added as well.

This change uses thread-local storage in order to get access to the endpoint in the PJSIP
callback. If this is deemed usable, then I may not need to backport the changes I pulled
in from PJSIP trunk early in this branch's lifetime.