Convert to using string fields for channel snapshots
Change channel snapshots over to string fields and create copy,
comparison, and destructor functions for them (and using those
everywhere).

The event system now has support for handling blobs that contain
pointers and need use of these types of functions. This includes the
addition of the new PLTYPE RAW_LOCAL which stores the copy, cmp, and
destroy functions in the ie_maps array and usage of those functions in
event duplication and destruction functions.

Fix segfaults in the manager cache hashing and comparison functions
when OBJ_KEY is used.

Fix bugs in manager event generation from channel state updates.