Add initial module refcounting and fix a crash when executing "core stop now".
If a module is loaded which uses either res_sip or res_sip_session the reference count of
either module will be increased to prevent unloading.

This also fixes a crash issue with the OPTIONS support. If res_sip were unloaded and then
"core stop now" executed an atexit function was still attempting to be called by the core
when it should not have been. The OPTIONS support now uses the SIP module start() and stop()
callbacks to set up and tear down resources.