Use the threadpool serializer.
Instead of an ast_sip_work structure, we now just make
use of the ast_threadpool_serializer() API and store an
ast_taskprocessor. This helps to reduce code significantly
and using general-purpose APIs is a good thing.

In addition, I changed the transport state shutdown not to
create a serializer since it isn't necessary.