Add a distributor for inbound requests and responses.
This distributor sits underneath the PJSIP transaction layer
and moves all incoming requests and responses into the
SIP threadpool.

The purpose behind this is that we will soon be placing endpoint
identification and authentication at this same point so that
applications can grab the endpoint from the rdata and not have
to worry about authentication.

The distributor is at this part of the stack in order to ease
some complexity in processing. The distributor simply has to
be called into by the PJSIP endpoint. This prevents us from
having to also register the distributor as a transaction or
dialog user.

The reason we want to distribute the incoming messages into the
threadpool is that endpoint lookup and authentication can each
potentially take a (relatively) long time to complete if something
like a remote database lookup is part of the process. For long-running
tasks, we do not want to hold up incoming request processing by the
PJSIP endpoint, so we farm the task off to the threadpool instead.

The next step will be to add the endpoint identifier and authentication
process.