Change to only authenticating dialog-creating methods.
This helps us not to try to authenticate things like BYE and INFO, but
it doesn't really help with other in-dialog requests like reinvites. We'll
still challenge those.

We unfortunately cannot detect if a request is in-dialog since the request
has not reached the dialog layer before it reaches us.