Streamline the process for new session creation.
On incoming out-of-dialog INVITEs, the on_rx_request()
callback is now called in our threadpool. This means there is
no reason to push the handling of the INVITE to the threadpool.
Now we just handle it in-line. This means that there is no good
reason to pass a serializer into ast_sip_session_alloc(). Instead,
the session can just create its own serializer during allocation.

I also changed the behavior of ast_sip_session_alloc() to increase
the refcount of the passed-in endpoint by one since it is keeping
a reference to the endpoint. This leads to much cleaner code that
calls ast_sip_session_alloc() since it can always unconditionally
unref its local endpoint.

With these two changes in place, the somewhat awkward destroy_serializer
and destroy_endpoint booleans are now gone from incoming INVITE
handling.