I finally got some funding for continuing my work on RTCP support in Asterisk and bringing it closer to trunk, starting with release 1.8.
I know parts of this code is messy so feedback on improvements are always welcome,
as is fellow coders. 

The goal is to be able to produce CQRs, Call Quality Records, for every SIP call leg
and to be able to watch call quality in realtime with AMI. The current 1.4 branch
can store these records in mysql - I have a few million of them in a production
server.

If someone has the knowledge and time to add RTCP XR it would be awsome.