Fix the memory corruption/void* lifetime management problems
Change the definition of the callback to include the
stasis_subscription object so that the uniqueid can be checked against
incoming messages to appropriately manage the lifetime of objects
handed to the subscription that are in need of destruction.

Correct stasis_cache to manage object lifetime properly.

Correct stasis tests to manage object lifetime properly.

Introduce a new function (stasis_subscription_final_message) to assist
in determining which message is the callback's final message.