Remove the session_on_tsx_state callback.
This callback could be useful if we expected to handle transactions
that were out-of-dialog in the session module. However, we purposely
have designed our implementation not to work that way, so there is
no need for this function to exist.

Instead, we have the inv_session-specific transaction state callback
we use instead. It's mostly the same except that we are given a handle
to the inv_session as part of the function parameters.