Resurrect res_sip_endpoint_identifier_ip and make it configurable.
By configurable I mean you now specify in res_sip.conf specific addresses (or ranges)
that should match and to what endpoint. This is more flexible in that you can have a
single ITSP with multiple IPs all match back to a single endpoint, so less hassle in the end.